Move Tendencies class into separate files
