+++ components/omega/src/timeStepping/TimeStepper.h
@@ -15,7 +15,7 @@
 #include "Halo.h"
 #include "HorzMesh.h"
 #include "OceanState.h"
-#include "TendencyTerms.h"
+#include "Tendencies.h"
 #include "TimeMgr.h"
 #include "Tracers.h"
